Добавлена демонстрация окраски
Добавлена демонстрация окраски графов в белый/черный цвет. Вывод при компиляции:
Coloring Result:
Vertex 1: Black
Vertex 2: White
Vertex 3: White
Vertex 4: White
Vertex 5: White
Vertex 6: White
