+++ list7.15.cpp
@@ -1,5 +1,6 @@
 #include <stdbool.h>
-#include <stdio.h> // добавим подключение заголовочного файла для функции printf
+#include <stdio.h>
+#include <stdlib.h>
 
 #define MAXV 1000
 
@@ -9,14 +10,14 @@
 
 typedef struct {
     int nvertices;
-    // другие поля графа
+    int adj[MAXV + 1][MAXV + 1]; // adjacency matrix
 } graph;
 
 bool bipartite;
-int color[MAXV+1];
-bool discovered[MAXV+1];
+int color[MAXV + 1];
+bool discovered[MAXV + 1];
 
-// объявление функций
+// function declarations
 void initialize_search(graph *g);
 void bfs(graph *g, int start);
 
@@ -36,28 +37,77 @@ void twocolor(graph *g) {
     }
 }
 
-// другие функции оставляем без изменений
-void process_edge(int x, int y);
-int complement(int color);
+// function definitions
+void initialize_search(graph *g) {
+    int i;
+    for (i = 1; i <= g->nvertices; i++) {
+        discovered[i] = false;
+    }
+}
+
+void bfs(graph *g, int start) {
+    int queue[MAXV]; // queue for BFS traversal
+    int front = 0, rear = 0; // front and rear pointers for the queue
+
+    discovered[start] = true;
+    queue[rear++] = start; // enqueue the starting vertex
+
+    while (front < rear) {
+        int v = queue[front++]; // dequeue a vertex from the queue
+
+        // process the vertex v (color it)
+        // Here, we color the vertex v with the opposite color of its parent
+        int parentColor = color[v];
+        int vertexColor = (parentColor == WHITE) ? BLACK : WHITE;
+        color[v] = vertexColor;
+
+        // process all the adjacent vertices of v
+        for (int u = 1; u <= g->nvertices; u++) {
+            if (g->adj[v][u]) { // there is an edge between v and u
+                if (!discovered[u]) {
+                    discovered[u] = true;
+                    queue[rear++] = u; // enqueue the adjacent vertex
+                }
+            }
+        }
+    }
+}
+
+void process_edge(int x, int y) {
+    // implementation of process_edge function
+}
+
+int complement(int color) {
+    // implementation of complement function
+}
+
 int main() {
     graph g;
     g.nvertices = 6;
-    // инициализация графа, добавление ребер и прочие действия...
+    // initialize the graph, add edges, and perform other actions...
+    // For example, let's say we have the following edges:
+    // 1-2, 1-3, 2-4, 3-5, 4-5, 4-6
+    g.adj[1][2] = g.adj[2][1] = 1;
+    g.adj[1][3] = g.adj[3][1] = 1;
+    g.adj[2][4] = g.adj[4][2] = 1;
+    g.adj[3][5] = g.adj[5][3] = 1;
+    g.adj[4][5] = g.adj[5][4] = 1;
+    g.adj[4][6] = g.adj[6][4] = 1;
 
     twocolor(&g);
 
-    // Выводим результат раскраски
+    // Print the coloring result
     int i;
-    printf("Результат раскраски:\n");
+    printf("Coloring Result:\n");
     for (i = 1; i <= g.nvertices; i++) {
         if (color[i] == WHITE) {
-            printf("Вершина %d: Белый\n", i);
+            printf("Vertex %d: White\n", i);
         } else if (color[i] == BLACK) {
-            printf("Вершина %d: Черный\n", i);
+            printf("Vertex %d: Black\n", i);
         } else {
-            printf("Вершина %d: Не раскрашена\n", i);
+            printf("Vertex %d: Not colored\n", i);
         }
     }
 
     return 0;
-}
+}
